funcionalidad delc carrito ok

+++ database/migrations/2025_10_23_215541_add_product_lot_id_to_stock_adjustments_table.php
@@ -12,7 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('stock_adjustments', function (Blueprint $table) {
-            $table->foreignId('product_lot_id')->nullable()->constrained('product_lots')->onDelete('set null')->after('location_id');
+            if (!Schema::hasColumn('stock_adjustments', 'product_lot_id')) {
+                $table->foreignId('product_lot_id')->nullable()->constrained('product_lots')->onDelete('set null')->after('location_id');
+            }
         });
     }
